intf/src/main/java/com/stubhub/domain/inventory/v2/listings/intf/DeliveryMethod.java
package com.stubhub.domain.inventory.v2.listings.intf;


import javax.xml.bind.annotation.XmlAccessType;
import javax.xml.bind.annotation.XmlAccessorType;
import javax.xml.bind.annotation.XmlRootElement;
import javax.xml.bind.annotation.XmlType;

@XmlAccessorType(XmlAccessType.FIELD)
@XmlRootElement(name = "deliveryMethod")
@XmlType(name = "", propOrder = { "id","name", "estimatedDeliveryTime"})
public class DeliveryMethod {
	
	private Long id;
	private String name;
	private String estimatedDeliveryTime;
	
	public Long getId() {
		return id;
	}
	public void setId(Long id) {
		this.id = id;
	}
	public String getName() {
		return name;
	}
	public void setName(String name) {
		this.name = name;
	}
	public String getEstimatedDeliveryTime() {
		return estimatedDeliveryTime;
	}
	public void setEstimatedDeliveryTime(String estimatedDeliveryTime) {
		this.estimatedDeliveryTime = estimatedDeliveryTime;
	}

}

